refactor(auth): extract email login form component
Move landing-page email auth form/dialog internals out of LoginButton into a dedicated EmailLoginForm component.
